backend_d/requirements.txt
asgiref==3.7.2
Django==4.2.10
django-allauth==0.57.0
django-cors-headers==4.6.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.4.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==4.1.0
freezegun==1.4.0
Pillow==10.1.0
pytest-xdist==3.5.0
django-phonenumber-field[phonenumbers]==7.3.0
google-auth==2.28.0
django-otp==1.3.0
twilio==8.13.0
cryptography==41.0.7
requests-mock==1.11.0
django-filter==23.5
gunicorn==21.2.0
dj-database-url==2.1.0
celery==5.3.6
redis==5.0.1  # Redis as message broker
django-celery-beat==2.5.0  # For periodic tasks